pyproject.toml
[project]
name = "anthropic-openrouter-proxy"
version = "0.1.0"
description = "A proxy server to convert between Anthropic and OpenAI API formats for OpenRouter"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.uv]
[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
